apps/inference/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
torch>=2.0.0
open_clip_torch>=2.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.28.0
